address review comment

+++ ratis-common/src/main/java/org/apache/ratis/util/SizeInBytes.java
@@ -56,11 +56,11 @@ public final class SizeInBytes {
     return new SizeInBytes(size, input, description);
   }
 
-  public static int byteToMb (int size){
+  public static int byteToMb (long size){
     if (size > 0 && size <  (ONE_MB.getSizeInt())) {
       return 1;
     }
-    return size/ ONE_MB.getSizeInt();
+    return (int) (size/ ONE_MB.getSizeInt());
   }
 
   private final long size;
